% Generated by roxygen2 (4.0.2): do not edit by hand
\name{BarData}
\alias{BarData}
\title{Function getting historical bar data}
\usage{
BarData(tickers = "AAPL US", type = "Equity",
  start.date.time = "2014-01-01 09:30:00",
  end.date.time = "2014-08-15 16:05:00", time.zone = "America/New_York",
  interval = "5")
}
\arguments{
\item{tickers}{character vector specifying the tickers}

\item{type}{a character string stating the security class
such as equities "Equity" or currencies "Curncy". The input is not
case sensitive}

\item{start.date.time}{a character string specifying the start date
and time}

\item{end.date.time}{a character string specifying the end date
and time}

\item{time.zone}{which time zone should be attached to the
specified date and time}

\item{interval}{a character string stating the bar interval
in minutes. One minute is the highest frequency allowed}
}
\description{
Function getting historical bar data
}
\details{
In the current version the function has only traded bars as
its default input. Bid and ask bar data points are possible and
will be added in the future.

The time zone input is added to make it easier to set times as the
local time for when the exchange is open never changes e.g. always
09:30 to 16:00 for NYSE and NASDAQ unless on a holiday, whereas
GMT changes two times a year because of daylight saving.

The data output is a xts object. For multiple tickers the xts
objects are stored in a list as missing bar data can in some
cases create unequal length of bar data for each ticker making
it impossible to wrap the data into an array.

As especially stock prices are impacted on ex-dividend dates
the option "adjustmentFollowDPDF" is to TRUE as default in
data query to Bloomberg. The DPDF function in the Bloomberg
terminal is where the user sets "Dividend & Corporate Action
Settings", so if the user has selected adjustment for normal
and abnormal cash dividends (splits) then these adjustments
will be reflected in the output bar data.
}
\references{
The official time zone names accepted for POSIX are found
here:
\url{http://en.wikipedia.org/wiki/List_of_tz_database_time_zones}
}

