% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bar_point.R
\name{BarPoint}
\alias{BarPoint}
\title{Function extracting the intraday price based on input date and time
on input date and time variable}
\usage{
BarPoint(data, bar.date = "2014-04-14", bar.time = "19:58:00",
  bar.type = "close")
}
\arguments{
\item{data}{a list of market bar data in the format created
from the bar data function}

\item{bar.date}{a character string specifying the date for bar
price point extraction}

\item{bar.time}{a character string specifying the time for bar
price point extraction (see details)}

\item{bar.type}{a character string specifying the price point e.g.
open, high, low, close or volume}
}
\description{
Function extracting the intraday price based on input date and time
on input date and time variable
}

