% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data_point.R
\name{DataPoint}
\alias{DataPoint}
\title{Function getting a single data point}
\usage{
DataPoint(tickers = "ADS GY", type = "Equity", fields = "PX_LAST",
  override.fields = NULL, override.values = NULL)
}
\arguments{
\item{tickers}{character vector specifying the securities}

\item{type}{a character string specifying instrument type
price point extraction}

\item{fields}{character vector specifying data fields
price point extraction (see details)}

\item{override_fields}{character vector specifying override fields}

\item{override_values}{character vector specifying override values}
}
\description{
Function getting a single data point
}

