% Generated by roxygen2 (4.0.2): do not edit by hand
\name{HistIndexTickers}
\alias{HistIndexTickers}
\title{Get historical index tickers to avoid surviourship bias}
\usage{
HistIndexTickers(index = "SPX", field = "ticker", startdate = "20140101",
  enddate = "", freq = "MONTHLY")
}
\arguments{
\item{index}{a character string specifying the index ticker}

\item{field}{specifying the output instrument identification}

\item{startdate}{a character string specifying the start date}

\item{enddate}{a character string specifying the end date}

\item{freq}{a character string specifying time frequency}
}
\description{
The function works for one equity index at a time
}
\details{
The field variable has two options: "ticker" or "isin"
enabling the function to return the desired instrument identification.

The frequency variable is set using the standard Bloomberg inputs such
as "MONTHLY", "DAILY" etc.
}

