% Generated by roxygen2 (4.0.2): do not edit by hand
\name{IndexMembers}
\alias{IndexMembers}
\title{Function loading all tickers of a specified index}
\usage{
IndexMembers(index = "SPX")
}
\arguments{
\item{index}{Bloomberg index ticker code}
}
\description{
Function loading all tickers of a specified index
}
\examples{
IndexMembers(index = "SPX")
}

