% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/market_data.R
\name{MarketData}
\alias{MarketData}
\title{Function that returns market data (Open, High, Low, Close, Volume)}
\usage{
MarketData(tickers = "AAPL US", type = "Equity", freq = "DAILY",
  currency = NULL, startdate = "20140101", enddate = "",
  non.trading.days = "NON_TRADING_WEEKDAYS",
  non.trading.days.fill = "NIL_VALUE", follow.dpdf = TRUE)
}
\arguments{
\item{tickers}{Character vector of ticker names (see details)}

\item{type}{Character that specifying instrument type (see details)}
}
\description{
Function that returns market data (Open, High, Low, Close, Volume)
}
\details{
The function returns conventional market data such as open,
high, low, close, volume. The output object is either a xts
object or array depending on the number of tickers provided (see details).

Tickers are structured as symbols together with the exchange code e.g.
AAPL US for Apple across aggregate US exchanges (composite code) and
AAPL UW for Apple's price data from NASDAQ Global Select
}
\examples{
MarketData(index = "AAPL US", startdate = "20140101", enddate = "20140630")
}

