% Generated by roxygen2 (4.0.2): do not edit by hand
\name{PointInTime}
\alias{PointInTime}
\title{Function getting historical point-in-time fundamental data}
\usage{
PointInTime(tickers = "AAPL US", type = "Equity",
  fields = "TRAIL_12M_EPS", currency = NULL, startdate = "20050101",
  enddate = "")
}
\arguments{
\item{tickers}{character vector specifying the securities}

\item{type}{a character string specifying instrument type
price point extraction}

\item{fields}{character vector specifying data fields
price point extraction (see details)}

\item{override_fields}{character vector specifying override fields}

\item{override_values}{character vector specifying override values}
}
\description{
Function getting historical point-in-time fundamental data
}

