% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/vwap_point.R
\name{VwapPoint}
\alias{VwapPoint}
\title{Function calculating intraday VWAP and extract a VWAP value based
on input date and time variable}
\usage{
VwapPoint(data, vwap.date = "2014-04-14", vwap.time = "19:59:00")
}
\arguments{
\item{data}{a list of market bar data in the format created
from the bar data function}

\item{vwap.date}{a character string specifying the date for VWAP
price point extraction}

\item{vwap.time}{a character string specifying the time for VWAP
price point extraction (see details)}
}
\description{
Function calculating intraday VWAP and extract a VWAP value based
on input date and time variable
}

