% Generated by roxygen2 (4.0.2): do not edit by hand
\name{BarData}
\alias{BarData}
\title{Get historical bar data on financial instruments}
\usage{
BarData(tickers = "AAPL US", type = "Equity", start.date = "2014-01-01",
  start.time = "09:30:00", end.date = NULL, end.time = NULL,
  time.zone = "America/New_York", interval = "5")
}
\arguments{
\item{tickers}{Character vector specifying the ticker code(s)}

\item{type}{Specify security class such as equities "Equity" or
currencies "Curncy". The input is not case sensitive.}

\item{start.date}{A character vector specifying the start date in
date class format. See details.}

\item{start.time}{A character vector specifying the start time.
See details.}

\item{end.date}{A character vector specifying the start date in
data class format.}

\item{end.time}{A character vector.}

\item{time.zone}{The time zone should be attached to the
specified date and time. See details.}

\item{interval}{A character vector stating the bar interval in
minutes. One minute is the highest frequency allowed.}
}
\description{
The function works for one or multiple tickers across
many security classes
}
\details{
In the current version the function has only traded bars as
its default input. Bid and ask bar data points are possible and
will be added in the future.

The date inputs should follow R's Date Class format "\%Y-\%m-\%d" e.g.
"2014-01-01" for January 1, 2014. The end date's default value is NULL
which the function interprets as today. Time inputs should the standard
POSIX format "\%H:\%M:\%S" e.g. "09:30:00".

The time zone input is added to make it easier to set times, as the
local time for when the exchange is open never changes, e.g. always
09:30 to 16:00 for NYSE and NASDAQ unless on a holiday, whereas
GMT changes two times a year because of daylight saving.

The data output is a xts object. For multiple tickers the xts
objects are stored in a list as missing bar data can in some
cases create unequal length of bar data for each ticker making
it difficult to wrap the data into an array. It becomes even more
cumbersome for European equities. The attached time zone is set to
the system's time zone.

As especially stock prices are impacted on ex-dividend dates
the option "adjustmentFollowDPDF" is to TRUE as default in
data query to Bloomberg. The DPDF function in the Bloomberg
terminal is where the user sets "Dividend & Corporate Action
Settings", so if the user has selected adjustment for normal
and abnormal cash dividends (splits) then these adjustments
will be reflected in the output bar data.
}
\examples{
BarData()  # Returns six months of 5 minutes bar data on Apple
}
\references{
The official time zone names accepted for POSIX are found
here:
\url{http://en.wikipedia.org/wiki/List_of_tz_database_time_zones}
}

