% Generated by roxygen2 (4.0.2): do not edit by hand
\name{HistData}
\alias{HistData}
\title{Function getting historical Bloomberg data}
\usage{
HistData(tickers = "GS US", type = "Equity", fields = "PX_LAST",
  freq = "DAILY", startdate = "20140101", enddate = "", currency = NULL,
  non.trading.days = "NON_TRADING_WEEKDAYS",
  non.trading.days.fill = "NIL_VALUE", calendar.type = "CALENDAR",
  pricing = "PRICING_OPTION_PRICE", follow.dpdf = TRUE,
  override.fields = NULL, override.values = NULL)
}
\arguments{
\item{tickers}{character vector specifying the securities}

\item{type}{a character string specifying instrument type
price point extraction}

\item{fields}{character vector specifying data fields
price point extraction (see details)}

\item{freq}{a character string specifying time frequency}

\item{currency}{a character string specifying the currency}
}
\description{
Function getting historical Bloomberg data
}

