% Generated by roxygen2 (4.0.2): do not edit by hand
\name{HistEarnDates}
\alias{HistEarnDates}
\title{Get historical earnings dates across stocks}
\usage{
HistEarnDates(tickers = "JPM US")
}
\arguments{
\item{tickers}{Character vector specifying the ticker code(s)}
}
\description{
The function works for one or multiple tickers across the
equity class
}
\details{
For a single ticker the function returns a data
frame containing historical earnings release dates and the
reporting period.

For multiple tickers the function returns a list of data
frames containing the same information.
}

