% Generated by roxygen2 (4.0.2): do not edit by hand
\name{HistEcoDates}
\alias{HistEcoDates}
\title{Function getting historical Bloomberg data}
\usage{
HistEcoDates(tickers = "EURR002W", startdate = "20000101")
}
\arguments{
\item{tickers}{character vector specifying the securities}

\item{startdate}{numeric specified as "20000101" for "2000-01-01"}
}
\description{
Function getting historical Bloomberg data
}

