% Generated by roxygen2 (4.0.2): do not edit by hand
\name{HistIndexTickers}
\alias{HistIndexTickers}
\title{Function getting historical index tickers to avoid surviourship bias}
\usage{
HistIndexTickers(index = "SPX", startdate = "20090101", enddate = "",
  freq = "MONTHLY")
}
\arguments{
\item{index}{a character string specifying the index}

\item{startdate}{a character string specifying the start date}

\item{enddate}{a character string specifying the end date}

\item{freq}{a character string specifying time frequency}
}
\description{
Function getting historical index tickers to avoid surviourship bias
}

