% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/point_in_time.R
\name{PointInTime}
\alias{PointInTime}
\title{Function getting historical point-in-time fundamental data (beta version)}
\usage{
PointInTime(tickers = "AAPL US", type = "Equity",
  fields = "SALES_REV_TURN", currency = NULL, startdate = "20090101",
  enddate = "", sh.out = TRUE)
}
\arguments{
\item{tickers}{character vector specifying the securities}

\item{type}{a character string specifying instrument type
price point extraction}

\item{fields}{character vector specifying data fields
price point extraction (see details)}

\item{override_fields}{character vector specifying override fields}

\item{override_values}{character vector specifying override values}
}
\description{
Function getting historical point-in-time fundamental data (beta version)
}

